--- src/Ratatoskr.EfCore/Exceptions/OutboxMessageSerializationException.cs
@@ -1,7 +1,0 @@
-﻿namespace Ratatoskr.EfCore;
-
-public class OutboxMessageSerializationException(string message, string serializedProperties)
-    : OutboxException(message)
-{
-    public string SerializedProperties { get; } = serializedProperties;
-}
