+++ src/Ratatoskr.EfCore/Internal/OutboxMessageEntity.cs
@@ -4,16 +4,9 @@ using Ratatoskr.Core;
 
 namespace Ratatoskr.EfCore.Internal;
 
-internal class OutboxMessageEntity
+internal class OutboxMessageEntity : BaseMessageEntity
 {
     public Guid Id { get; private set; }
-    
-    public required byte[] Content { get; init; }
-
-    /// <summary>
-    /// JSON serialized properties
-    /// </summary>
-    public required string SerializedProperties { get; init; }
     
     public required DateTimeOffset CreatedAt { get; init; }
     
@@ -43,12 +36,6 @@ internal class OutboxMessageEntity
     public DateTimeOffset? ProcessingStartedAt { get; private set; }
 
     /// <summary>
-    /// The transport this outbox entry targets (e.g. "rabbitmq", "efcore").
-    /// </summary>
-    [MaxLength(50)]
-    public string TransportName { get; private set; } = string.Empty;
-
-    /// <summary>
     /// Optimistic concurrency token. Incremented on every state mutation to prevent
     /// two concurrent processors from processing the same message.
     /// </summary>
@@ -58,12 +45,6 @@ internal class OutboxMessageEntity
     /// Counts how many times this message has been requeued via the management API.
     /// </summary>
     public int RequeuedCount { get; private set; }
-
-    private MessageProperties? _cachedProperties;
-
-    public MessageProperties GetProperties() =>
-        _cachedProperties ??= JsonSerializer.Deserialize<MessageProperties>(SerializedProperties)
-        ?? throw new OutboxMessageSerializationException("Could not deserialize the message properties.", SerializedProperties);
 
     private OutboxMessageEntity(){}
     public static OutboxMessageEntity Create(byte[] message, MessageProperties props, TimeProvider timeProvider, string transportName)
