+++ tests/Ratatoskr.Tests/Integration/Management/ManagementCoverageTests.cs
@@ -6,8 +6,8 @@ using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.DependencyInjection;
 using Ratatoskr.EfCore.Internal;
 using Ratatoskr.EfCore.Management;
-using Ratatoskr.Endpoints;
 using Ratatoskr.Management;
+using Ratatoskr.Management.Endpoints;
 using Ratatoskr.Tests.Fixtures;
 
 namespace Ratatoskr.Tests.Integration.Management;
